% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lzstringr-package.R
\name{compressToBase64}
\alias{compressToBase64}
\title{Compress a string to Base64}
\usage{
compressToBase64(string)
}
\arguments{
\item{string}{A character string to be compressed.}
}
\value{
A character string representing the compressed input string in Base64 format.
}
\description{
This function takes a string as input and returns a compressed version of the string in Base64 format.
}
\examples{
\dontrun{
compressToBase64("Hello, world!")
}
}
