% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lzstringr-package.R
\name{compressToEncodedURIComponent}
\alias{compressToEncodedURIComponent}
\title{Compress a string to Encoded URI Component}
\usage{
compressToEncodedURIComponent(string)
}
\arguments{
\item{string}{A character string to be compressed.}
}
\value{
A character string representing the compressed input string in Encoded URI Component format.
}
\description{
This function takes a string as input and returns a compressed version of the string in Encoded URI Component format.
}
\examples{
compressToEncodedURIComponent("Hello, world!")
}
