% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lzstringr-package.R
\name{decompressFromBase64}
\alias{decompressFromBase64}
\title{Decompress a string from Base64}
\usage{
decompressFromBase64(string)
}
\arguments{
\item{string}{A character string in Base64 format to be decompressed.}
}
\value{
A character string representing the decompressed input string.
}
\description{
This function takes a compressed string in Base64 format as input and returns the decompressed version of the string.
}
\examples{
x <- compressToBase64("Hello, world!")
decompressFromBase64(x)
}
