% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lzstringr-package.R
\name{decompressFromEncodedURIComponent}
\alias{decompressFromEncodedURIComponent}
\title{Decompress a string from Encoded URI Component}
\usage{
decompressFromEncodedURIComponent(string)
}
\arguments{
\item{string}{A character string in Encoded URI Component format to be decompressed.}
}
\value{
A character string representing the decompressed input string.
}
\description{
This function takes a compressed string in Encoded URI Component format as input and returns the decompressed version of the string.
}
\examples{
x <- compressToEncodedURIComponent("Hello, world!")
decompressFromEncodedURIComponent(x)
}
